Create a dedicated user for subllime-text image.
This allows us to finally run a GUI from Docker though this is by sheer
coincidence.

Figured out that sharing $DISPLAY and /tmp/.X11-unix is not working because of
uid bits.

In most of my machines, the user I am using is the user that was first created
on this machine. This fixes our problem because creating a dedicated user in
the Docker image means this user gets the same uid as the one I am using on the
host machine. However, this means that this trick will not work when ran as
other users. One solution is to create a build arg for base that will set a
user for all these images with the same uid as the host machine user.