Mount the home directory into the app container.
This is more justification for not giving a punny name to the user in
the base Docker container.

The idea here is that the apps that should be dockerized like this are
the ones who only write to the home directory. If they need to write
elsewhere to function, you probably shouldn't be running them with
Docker.